iterations/4-haystack-mcp/requirements.txt
haystack-ai
python-dotenv
"mcp[cli]"
mcp-haystack
